Final_version
Updated all the scripts with errors / bugs / poor coding that we encountered during master students projects and own work. Functionality for QligFEP and QresFEP are updated from Alejandro's and Lucien's improvements. (Final versions for proper dual topology FEP implementation to follow).
This should be the template version of QligFEP for new students joining the group.

Co-Authored-By: Alejandro Díaz Holguín <56196173+shevsea@users.noreply.github.com>